Move `Bytes` to `read::util`
This also means we don't need `alloc` when no features are enabled.
