+++ src/lib.rs
@@ -64,6 +64,7 @@
 #[cfg(feature = "cargo-all")]
 compile_error!("'--all-features' is not supported; use '--features all' instead");
 
+#[cfg(feature = "read_core")]
 #[allow(unused_imports)]
 #[macro_use]
 extern crate alloc;
