+++ src/read/mod.rs
@@ -5,7 +5,6 @@ use alloc::vec::Vec;
 use core::{fmt, result};
 
 use crate::common::*;
-use crate::ByteString;
 
 mod read_ref;
 pub use read_ref::*;
@@ -16,7 +15,7 @@ mod read_cache;
 pub use read_cache::*;
 
 mod util;
-pub use util::StringTable;
+pub use util::*;
 
 #[cfg(any(
     feature = "coff",
